requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
attrs==23.1.0
    # via
    #   automat
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
black==23.10.0
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via cryptography
channels[daphne]==4.0.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.1.0
    # via -r requirements.in
charset-normalizer==3.3.0
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
constantly==15.1.0
    # via twisted
cryptography==41.0.4
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
    #   social-auth-core
daphne==4.0.0
    # via
    #   -r requirements.in
    #   channels
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.6
    # via
    #   -r requirements.in
    #   channels
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   social-auth-app-django
django-environ==0.11.2
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.0
    # via djoser
djoser==2.2.0
    # via -r requirements.in
drf-yasg[validation]==1.21.7
    # via -r requirements.in
environ==1.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==22.10.0
    # via twisted
inflection==0.5.1
    # via drf-yasg
jsonschema==4.19.1
    # via swagger-spec-validator
jsonschema-specifications==2023.7.1
    # via jsonschema
msgpack==1.0.7
    # via channels-redis
mypy-extensions==1.0.0
    # via black
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==23.2
    # via
    #   black
    #   build
    #   drf-yasg
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.11.0
    # via black
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyopenssl==23.2.0
    # via twisted
pyproject-hooks==1.0.0
    # via build
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   drf-yasg
    #   swagger-spec-validator
redis==5.0.1
    # via channels-redis
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rpds-py==0.10.6
    # via
    #   jsonschema
    #   referencing
service-identity==23.1.0
    # via twisted
six==1.16.0
    # via automat
social-auth-app-django==5.4.0
    # via djoser
social-auth-core==4.4.2
    # via social-auth-app-django
sqlparse==0.4.4
    # via django
swagger-spec-validator==3.0.3
    # via drf-yasg
twisted[tls]==23.8.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.8.0
    # via
    #   swagger-spec-validator
    #   twisted
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.0.7
    # via requests
wheel==0.41.2
    # via pip-tools
zope-interface==6.1
    # via
    #   -r requirements.in
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
